requirements.txt
shapely==1.7.0
tensorflow-gpu==1.15.0
h5py==2.10.0
tifffile==0.15.1
numpy
pandas==1.0.5
opencv==4.2.0
tqdm
scikit-image==0.16.2
pillow==7.1.2
matplotlib==3.2.2
jupyter
